feat: Refresh when .jj is touched -- UNSTABLE
This makes it possible to keep lazyjj in one window, and working
with jj in another. Every jj command will cause lazyjj to update.


UNSTABLE

Sometimes an external and an internal jj operation runs at the same time.
This causes duplicate branches to appear.

On the positive side, it is excellent for testing concurrency issues with jj.


TRY: Add a --no-workspace-update flag so the refresh can run
without updating .jj.

TRY: Let focus lost/gained control when updates are allowed.