feat: Refresh when .jj is touched -- TODO
This makes it possible to keep lazyjj in one window, and working
with jj in another. Every jj command will cause lazyjj to update.


TODO: Get rid of delay?

RwLock block file notifications while main thread is not waiting for events.
The idea is that any operation would modify jj, so we cannot trust any notification
before the application is idle.
This might also get rid of the delay. Now lazyjj will wake up immediately if left on its own.