fix: Keep scroll position during refresh
If you scroll down a document and then resize the
terminal window, the scroll will be reset to 0.
This is because lazyjj does a refresh.
This fix makes sure that does not happen.

You can reset scroll on right panel by changing
selection on left panel.

There is no way to reset scroll on left panel.
Maybe Home+End?